+++ examples/example_oilspill_seafloor.py
@@ -15,30 +15,36 @@ reader_norkyst = reader_netCDF_CF_generic.Reader(o.test_data_folder() + '14Jan20
 #reader_norkyst = reader_netCDF_CF_generic.Reader('https://thredds.met.no/thredds/dodsC/sea/norkyst800m/1h/aggregate_be')
 
 o.add_reader([reader_norkyst])
-o.fallback_values['x_wind'] = 0
-o.fallback_values['y_wind'] = 0
+o.fallback_values['x_wind'] = 3
+o.fallback_values['y_wind'] = 7
+o.set_config('drift:vertical_mixing', True)
 
 #%%
 # Seeding some particles
-lon = 4.5; lat = 62.0
 time = [reader_norkyst.start_time,
         reader_norkyst.start_time + timedelta(hours=1)]
-#%%
-# Seed oil elements at defined position and time
-o.seed_elements(lon, lat, z='seafloor', radius=5, number=3000, time=time)
-
-o.set_config('drift:vertical_mixing', True)  # Otherwise also no updrift
+o.seed_elements(lon=4.5, lat=62.0, z='seafloor', radius=0, number=3000,
+                time=time, oiltype='*GENERIC DIESEL')
 
 #%%
-# Running model
-o.run(steps=12*2, time_step=300, time_step_output=300)
+# Setting the range of droplet sizes for the seafloor release
+o.set_config('seed:droplet_diameter_min_subsea', 0.0001)
+o.set_config('seed:droplet_diameter_max_subsea', 0.0005)
+
+#%%
+# Running model with a small timestep to resolve the boyant rising
+o.run(duration=timedelta(hours=2), time_step=60, time_step_output=60)
 
 #%%
 # Print and plot results
 print(o)
-o.animation_profile()
 
+o.animation_profile()
 #%%
 # .. image:: /gallery/animations/example_oilspill_seafloor_0.gif
+
+o.animate_vertical_distribution(bins=30)
+#%%
+# .. image:: /gallery/animations/example_oilspill_seafloor_1.gif
 
 o.plot_oil_budget()