+++ tests/integration/test_run.py
@@ -231,7 +231,7 @@ class TestRun(unittest.TestCase):
     #    o.run(steps=3)
     #    o.write_geotiff('geotiff.tif')
 
-    def test1_seed_single_point_over_time(self):
+    def test_seed_single_point_over_time(self):
         """Test a model run"""
         self.make_OceanDrift_object()
         self.o.seed_elements(2.0, 61.0, radius=0, number=9,
@@ -252,42 +252,6 @@ class TestRun(unittest.TestCase):
         self.assertEqual(self.o.num_elements_activated(), 5)
         self.assertEqual(self.o.num_elements_deactivated(), 0)
         self.assertEqual(self.o.num_elements_total(), 9)
-
-    def test2_seed_elements(self):
-        """Test a model run"""
-        self.make_OceanDrift_object()
-        self.o.seed_elements([2.0, 5.0, 3.0], [61.0, 60.0, 62.0],
-                             radius=0, number=9,
-                             time=[datetime(2015, 1, 1), datetime(2015, 1, 3)])
-
-        # Check that 6 elements are scheduled, but none seeded
-        self.assertEqual(self.o.num_elements_scheduled(), 9)
-        self.assertEqual(self.o.num_elements_active(), 0)
-        self.assertEqual(self.o.num_elements_activated(), 0)
-        self.assertEqual(self.o.num_elements_deactivated(), 0)
-        self.assertEqual(self.o.num_elements_total(), 9)
-        # Run simulation
-        self.o.run(steps=30, outfile='unittest.nc')
-        # Check that 1 element is deactivated (stranded),
-        # 1 yet not seeded and 4 active
-        self.assertEqual(self.o.num_elements_scheduled(), 4)
-        self.assertEqual(self.o.num_elements_active(), 4)
-        self.assertEqual(self.o.num_elements_activated(), 5)
-        self.assertEqual(self.o.num_elements_deactivated(), 1)
-        self.assertEqual(self.o.num_elements_total(), 9)
-
-    def test3_run_import(self):
-        """Import output file from previous test, and check elements"""
-        self.o = OceanDrift(loglevel=20)
-        self.o.io_import_file('unittest.nc')
-        self.assertEqual(self.o.num_elements_active(), 4)
-        self.assertEqual(self.o.num_elements_activated(), 5)
-        self.assertEqual(self.o.num_elements_deactivated(), 1)
-        self.assertEqual(self.o.num_elements_total(), 5)
-
-    def test4_cleaning(self):
-        """Cleaning up"""
-        os.remove('unittest.nc')
 
     def test_temporal_seed(self):
         self.o = OceanDrift(loglevel=20)
@@ -327,9 +291,53 @@ class TestRun(unittest.TestCase):
 
         o1.run(steps=20, time_step=300, time_step_output=1800,
                export_buffer_length=10, outfile='verticalmixing.nc')
-        self.assertAlmostEqual(o1.history['z'].min(), -38.1, 1)
+        self.assertAlmostEqual(o1.history['z'].min(), -38.7, 1)
         self.assertAlmostEqual(o1.history['z'].max(), 0.0, 1)
         os.remove('verticalmixing.nc')
+
+    def test_vertical_mixing_profiles(self):
+        # Testing an isolated mixing timestep
+
+        cases = [  # Some cases with expected outcome
+                {'vt': 0, 'K': 0, 'K_below': .01, 'T': 60, # No mixing
+                    'zmin': -10, 'zmax': -10, 'zmean': -10},
+                {'vt': -.005, 'K': 0, 'K_below': .01, 'T': 60, # Sinking
+                    'zmin': -74.0, 'zmax': -21.4, 'zmean': -50.2},
+                {'vt': 0, 'K': .01, 'K_below': .01, 'T': 60, # Mixing
+                    'zmin': -39.8, 'zmax': -0.1, 'zmean': -14.5},
+                {'vt': .005, 'K': .01, 'K_below': .01, 'T': 60, # Mixing and rising
+                    'zmin': -8.1, 'zmax': -0.01, 'zmean': -2.1},
+                {'vt': -0.005, 'K': .01, 'K_below': .01, 'T': 60, # Mixing and sinking
+                    'zmin': -75.8, 'zmax': -20.7, 'zmean': -48.1},
+                {'vt': 0, 'K': .02, 'K_below': .001, 'T': 60,  # Mixing in mixed layer
+                    'zmin': -22.8, 'zmax': -0.1, 'zmean': -9.8},
+                ]
+
+        N=100
+        z = np.arange(0, -30, -2)
+        time = datetime.now()
+        for case in cases:
+            diffusivity = np.ones(z.shape)*case['K']
+            diffusivity[z<-15] = case['K_below']
+            o = OceanDrift(loglevel=20)
+            o.set_config('drift:vertical_mixing', True)
+            o.set_config('vertical_mixing:diffusivitymodel', 'environment')
+            o.set_config('vertical_mixing:timestep', case['T'])
+            o.seed_elements(lon=4, lat=60, z=-10, time=time, number=N,
+                            terminal_velocity=case['vt'])
+            o.time = time
+            o.time_step = timedelta(hours=2)
+            o.release_elements()
+            o.fallback_values['land_binary_mask'] = 0
+            o.environment = np.array(np.ones(N)*100,
+                            dtype=[('sea_floor_depth_below_sea_level',
+                                    np.float32)]).view(np.recarray)
+            o.environment_profiles = { 'z': z, 'ocean_vertical_diffusivity':
+                                      np.tile(diffusivity, (N, 1)).T}
+            o.vertical_mixing()
+            self.assertAlmostEqual(o.elements.z.min(), case['zmin'], 1)
+            self.assertAlmostEqual(o.elements.z.max(), case['zmax'], 1)
+            self.assertAlmostEqual(o.elements.z.mean(), case['zmean'], 1)
 
     def test_export_step_interval(self):
         # Export to file only at end